Fixes #1562 - Fix circular dependency
