+++ packages/clay-date-picker/src/WeekdayHeader.tsx
@@ -5,7 +5,7 @@
  */
 
 import React, {FunctionComponent, ReactNode} from 'react';
-import {firstDayOfWeek} from './ClayDatePicker';
+import {firstDayOfWeek} from './types';
 
 interface IChildrenProps {
 	key: number;
