+++ packages/clay-date-picker/src/index.tsx
@@ -25,28 +25,9 @@ import InputDate from './InputDate';
 import TimePicker from './TimePicker';
 import Weekday from './Weekday';
 import WeekdayHeader from './WeekdayHeader';
+import {firstDayOfWeek, IAriaLabels, IYears} from './types';
 
-export enum firstDayOfWeek {
-	Sunday = 0,
-	Monday = 1,
-	Tuesday = 2,
-	Wednesday = 3,
-	Thursday = 4,
-	Friday = 5,
-	Saturday = 6,
-}
-
-export interface IAriaLabels {
-	buttonDot: string;
-	buttonNextMonth: string;
-	buttonPreviousMonth: string;
-	input?: string;
-}
-
-export interface IYears {
-	end: number;
-	start: number;
-}
+export {firstDayOfWeek, IAriaLabels, IYears};
 
 interface Props {
 	/**
